Extract hook for assessment data
